requirements.txt
attrs==20.3.0
requests==2.24.0
pytest==6.2.4
